Create libstatspull_lazy
Add lazy loading support to libstatspull.

There are a few clients (surfaceflinger, bootanimation dependencies)
that run before the statsd apex is mounted and cannot link
libstatspull.so when the process starts. Lazy loading will load the
library when the apis to set/clear callbacks or create PullAtomMetadata
get called.

This change is modeled heavily off of aosp/1570701, which added
libnativehelper_lazy.

Bug: 178480419
Test: atest libstatspull_lazy_test

Change-Id: I9956f5a690aec8c7ff9f0834a3f36387e5d8b23f